pyproject.toml
[project]
name = "storage-tg-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.6"
dependencies = [
    "environs>=14.2.0",
    "python-telegram-bot>=22.3",
]
